Add all the merged files
