+++ requirements.txt
@@ -1,0 +1,20 @@
+astroid==2.0.4
+atomicwrites==1.2.1
+attrs==18.2.0
+Click==7.0
+Flask==1.0.2
+isort==4.3.4
+itsdangerous==1.1.0
+Jinja2==2.10
+lazy-object-proxy==1.3.1
+MarkupSafe==1.1.0
+mccabe==0.6.1
+more-itertools==4.3.0
+pluggy==0.8.0
+py==1.7.0
+pylint==2.1.1
+pytest==3.10.0
+six==1.11.0
+typed-ast==1.1.0
+Werkzeug==0.14.1
+wrapt==1.10.11
